Fixed a minor bug
When checking that a current part layout matches the hint, the score
calculation did not check that all variables (position, side, rotation)
were initialized. But without all of them, it's not possible to compute
the center position, which is how the hint is specified.
So it was possible to get a null pointer exception.